% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.handle_url}
\alias{.handle_url}
\title{Process the URL as usual or return query information}
\usage{
.handle_url(url, return_url = F, print_url = F)
}
\arguments{
\item{url}{NSR URL to get using RCurl}

\item{return_url}{If TRUE, returns the URL rather than the results of the getURL call}

\item{print_url}{If TRUE, the URL is printed in addition to returning the NSR results (or URL)}
}
\value{
If print_url = FALSE, NSR results from the specified query in JSON format.  IF print_url = T, the URL.
}
\description{
.handle_url returns the URL as usual, or else prints or returns the URL
}
\note{
Internal function for de-bugging.
}
\keyword{Internal}
