% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_simple.R
\name{.NSR_simple}
\alias{.NSR_simple}
\title{Check the native status for plant species in a political region}
\usage{
.NSR_simple(occurrence_dataframe, ...)
}
\arguments{
\item{occurrence_dataframe}{A properly formatted dataframe, see http://bien.nceas.ucsb.edu/bien/tools/nsr/batch-mode/}

\item{...}{Additional parameters passed to internal functions.}
}
\value{
Dataframe containing NSR results.
}
\description{
.NSR_simple returns information on native status for species within a political region.
}
\note{
This function is slower and less informative than the typical NSR. We recommend using the NSR function.
}
\examples{
\dontrun{

nsr_testfile <- 
read.csv("http://bien.nceas.ucsb.edu/bien/wp-content/uploads/2019/02/nsr_testfile.csv")

results <- .NSR_simple(occurrence_dataframe = nsr_testfile)

}
}
