% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_metadata.R
\name{NSR_metadata}
\alias{NSR_metadata}
\title{Get metadata on NSR sources}
\usage{
NSR_metadata()
}
\value{
data.frame containing NSR source metadata
}
\description{
NSR_metadata returns metadata on the sources used by the NSR.
}
\examples{
\dontrun{

nsr_sources <- NSR_metadata()
 
}
}
