% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsr_testfile}
\alias{nsr_testfile}
\title{Example NSR data}
\format{
A data.frame with 22 observations of 5 variables:
\describe{
  \item{taxon}{taxon, "Genus species" format, with specific epithet optional}
  \item{country}{country that a species occurs within}
  \item{state_province}{state/province that a species occurs within}
  \item{county_parish}{county/parish that a species occurs within}
  \item{user_id}{unique numeric ID that can be used to link to original dataset}
  ...
}
}
\source{
\url{http://biendata.org}
}
\usage{
nsr_testfile
}
\description{
A sample dataset showing the proper formatting of NSR inputs.
}
\keyword{datasets}
