% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR_version.R
\name{NSR_version}
\alias{NSR_version}
\title{Get metadata on current NSR version}
\usage{
NSR_version()
}
\value{
Dataframe containing current NSR version number, build date, and code version.
}
\description{
Return metadata about the current NSR version
}
\examples{
{
NSR_version_metadata <- NSR_version()
}

}
